fix(server): require auth for custom provider credential test (SSRF)
Move `testGenerativeModelCustomProviderCredentials` from `Query` to
`Mutation` and gate it with `IsNotReadOnly`, `IsNotViewer`, and
`IsAdminIfAuthEnabled`. The resolver makes outbound HTTP requests using
a user-supplied `baseUrl` and headers; as a query field it was
unauthenticated and reachable via introspection, making it usable to
SSRF cloud metadata endpoints (e.g. `metadata.google.internal`).

Add a regression test that asserts the field is not on `Query`, is on
`Mutation`, and carries the required permission classes. Update the
phoenix-server skill (`graphql-patterns.md`, `SKILL.md`) and `REVIEW.md`
with rules so side-effecting GraphQL fields cannot land on `Query` again.
